Add files via upload
In recent years, the popularity of social media has been contributing to disseminate information over internet. Due to exponential growth of information and lack of any central moderation, it is very hard to analyze and authenticate the credibility of information. The malicious users take an advantage and propagate fake news and rumors over in- ternet, which can create a panic situation among public, individuals and companies.